[APPWAVE] Release - V1.3 - Wirelesss,analog,controller full integration. Wireless task no testing, controller task no testing. FFT and MFCC libraries implementation for future development

+++ components/App/Application/Modules/ANA_APP/include/i_analog_data.h
@@ -24,7 +24,7 @@ typedef struct {
 } SBufferTime_t;
 
 typedef struct {
-    uint8_t data[BUFFER_SIZE];
+    uint32_t data[BUFFER_SIZE];
     size_t size;
     bool ready;
 } SDataBuffer_t;
