+++ components/App/Application/Modules/Internal_Comms/wireless_t_share.c
@@ -38,7 +38,7 @@ void wireless_controller_read(SWlsCtrlSensMsg_t *msg)
 void controller_wireless_send(
     SErrorInfo_t alarm,
     SSystemStatus_t status,
-    uint8_t *plant_signal,
+    uint32_t *plant_signal,
     SEnvData_t env_data,
     SPowerData_t power_data,
     SAxisData_t *axis_buff,
@@ -49,7 +49,7 @@ void controller_wireless_send(
         mutex_lock(CTRL_WLS_M_ID);
             memcpy(&_ctrl_msg._alarm, &alarm, sizeof(SErrorInfo_t));
             memcpy(&_ctrl_msg._status, &status, sizeof(SSystemStatus_t));
-            memcpy(_ctrl_msg._plant_signal, plant_signal,DATA_BUFFER_SIZE*sizeof(uint8_t));
+            memcpy(_ctrl_msg._plant_signal, plant_signal,DATA_BUFFER_SIZE*sizeof(uint32_t));
             memcpy(&_ctrl_msg._env_data, &env_data, sizeof(SEnvData_t));
             memcpy(&_ctrl_msg._power_data, &power_data, sizeof(SPowerData_t));
             memcpy(_ctrl_msg._axis_buff, axis_buff,DATA_BUFFER_SIZE*sizeof(SAxisData_t));
