+++ components/App/Application/Modules/WLS_APP/objects/objects.h
@@ -25,7 +25,7 @@
 const anjay_dm_object_def_t **plant_data_object_create(void);
 void custom_object_release(const anjay_dm_object_def_t **def);
 void custom_object_update(anjay_t *anjay,const anjay_dm_object_def_t *const *def);
-void plant_object_value_update(uint8_t plant_data,const anjay_dm_object_def_t *const *def);
+void plant_object_value_update(uint32_t plant_data,const anjay_dm_object_def_t *const *def);
 void plant_object_send(anjay_send_batch_builder_t *builder,anjay_t *anjay, const anjay_dm_object_def_t **def);
 /**************************************************************/
