+++ components/App/Application/Modules/WLS_APP/objects/test_obj.c
@@ -102,10 +102,10 @@ void custom_object_release(const anjay_dm_object_def_t **def) {
 
 }
 
-void plant_object_value_update(uint8_t plant_data,const anjay_dm_object_def_t *const *def)
+void plant_object_value_update(uint32_t plant_data,const anjay_dm_object_def_t *const *def)
 {
     custom_object_t *obj = get_obj(def);
-    obj->plant_value = (uint32_t)plant_data;
+    obj->plant_value = plant_data;
 }
 
 void plant_object_send(anjay_send_batch_builder_t *builder,anjay_t *anjay, const anjay_dm_object_def_t **def) {
